+++ packages/sdk/config/src/loaders/browser.js
@@ -8,10 +8,12 @@
 
 import fetch from 'node-fetch';
 
+import { FILE_DYNAMICS } from '../types';
+
 // Fix a bug making fetch not being properly bound with webpack.
 const fetchBound = fetch;
 
-const CONFIG_ENDPOINT = '/config/config.json';
+const CONFIG_ENDPOINT = `/config/${FILE_DYNAMICS}`;
 
 export const LocalStorage = (item = 'options') => {
   return JSON.parse(window.localStorage.getItem(item) || '{}');