+++ packages/sdk/react-client/src/containers/BotFactoryClientProvider.tsx
@@ -1,0 +1,50 @@
+//
+// Copyright 2022 DXOS.org
+//
+
+import React, { ReactNode, useEffect, useState } from 'react';
+
+import { BotFactoryClient } from '@dxos/bot-factory-client';
+import { PublicKey } from '@dxos/crypto';
+import { raise } from '@dxos/debug';
+
+import { BotFactoryClientContext, createBotFactoryClient, useConfig } from '../hooks';
+
+export interface BotFactoryClientProviderProps {
+  children?: ReactNode
+}
+
+/**
+ * BotFactoryClientProvider
+ */
+export const BotFactoryClientProvider = ({
+  children
+}: BotFactoryClientProviderProps) => {
+  const config = useConfig();
+  const [botFactoryClient, setBotFactoryClient] = useState<BotFactoryClient>();
+
+  useEffect(() => {
+    setImmediate(async () => {
+      const botFactoryClient = await createBotFactoryClient(config);
+      setBotFactoryClient(botFactoryClient);
+
+      // TODO(burdon): Rename property: 'runtime.services.bot-factory.topic'
+      const topic = config.get('runtime.services.bot.topic') ?? raise(new Error('Missing Bot factory topic.'));
+      await botFactoryClient.start(PublicKey.from(topic));
+    });
+
+    return () => {
+      void botFactoryClient?.stop();
+    };
+  }, []);
+
+  if (!botFactoryClient) {
+    return null;
+  }
+
+  return (
+    <BotFactoryClientContext.Provider value={botFactoryClient}>
+      {children}
+    </BotFactoryClientContext.Provider>
+  );
+};