+++ packages/sdk/react-framework/src/containers/index.ts
@@ -2,7 +2,7 @@
 // Copyright 2020 DXOS.org
 //
 
-export * from './Bot';
 export * from './ErrorBoundary';
+export * from './JoinDialog';
 export * from './RegistrySearch';
-export * from './Sharing';
+export * from './SharingDialog';