+++ packages/sdk/react-registry-client/src/containers/RegistryProvider.tsx
@@ -5,25 +5,30 @@
 import debug from 'debug';
 import React, { useState, useEffect, ReactNode } from 'react';
 
+import { ConfigProvider } from '@dxos/config';
 import { IRegistryClient } from '@dxos/registry-client';
 
 import { RegistryContext } from '../hooks';
-import { createRegistryContext, RegistryConfigProvider } from './config';
+import { createRegistryContext } from './config';
 
 const log = debug('dxos:react-registry-client:error');
 
 interface RegistryProviderProps {
-  children?: ReactNode
-  config?: RegistryConfigProvider
+  config?: ConfigProvider
   registry?: IRegistryClient
+  children?: ReactNode
 }
 
 /**
  * Initializes and provides a DXNS registry instance given a config object or config generator.
  * To be used with `useRegistry` hook.
  */
-export const RegistryProvider = ({ children, config = {}, registry }: RegistryProviderProps) => {
-  const [value, setValue] = useState<RegistryContext | undefined>(registry && { registry });
+export const RegistryProvider = ({
+  config = {},
+  registry,
+  children
+}: RegistryProviderProps) => {
+  const [context, setContext] = useState<RegistryContext | undefined>(registry && { registry });
   const [error, setError] = useState<undefined | Error>(undefined);
   if (error) {
     log(error);
@@ -34,7 +39,8 @@ export const RegistryProvider = ({ children, config = {}, registry }: RegistryPr
     if (!registry) {
       setImmediate(async () => {
         try {
-          setValue(await createRegistryContext(config));
+          const context = await createRegistryContext(config);
+          setContext(context);
         } catch (error: any) {
           setError(error);
         }
@@ -43,12 +49,12 @@ export const RegistryProvider = ({ children, config = {}, registry }: RegistryPr
   }, []);
 
   // Still loading.
-  if (!value) {
+  if (!context) {
     return null;
   }
 
   return (
-    <RegistryContext.Provider value={value}>
+    <RegistryContext.Provider value={context}>
       {children}
     </RegistryContext.Provider>
   );